requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.7.0
chromadb>=0.5.4
pypdf>=4.0.0
python-docx>=1.0.0
numpy>=1.26.0
